# Conflicts:
#	WireAuthentication/Sources/WireAuthentication/Components/VerificationCodeComponent.swift
#	WireAuthentication/Sources/WireAuthentication/Needle.generated.swift
#	WireAuthentication/Sources/WireAuthenticationUI/MockDependencies.swift
#	WireAuthentication/Sources/WireAuthenticationUI/Views/Login/LoginViaEmail/LoginViaEmailViewModel.swift
#	WireAuthentication/Sources/WireAuthenticationUI/Views/Login/VerificationCode/VerificationCodeViewModel.swift
#	WireAuthentication/Tests/WireAuthenticationUITests/LoginViaEmail/LoginViaEmailViewModelTests.swift
#	WireAuthentication/Tests/WireAuthenticationUITests/VerificationCode/VerificationCodeViewModelTests.swift

+++ wire-ios/Wire-iOS/Sources/UserInterface/Components/Constants.swift
@@ -21,13 +21,12 @@ import WireCommonComponents
 import WireDesign
 
 enum Constants {
-
     static var teamAccountViewImageInsets = UIEdgeInsets(top: 2, left: 2, bottom: 2, right: 2)
 }
 
-extension Float {
+extension CGFloat {
     enum ConversationButtonMessageCell {
-        static let verticalInset: Float = 8
+        static let verticalInset: CGFloat = 8
     }
 }
